+++ packages/patternfly-4/react-core/src/components/ApplicationLauncher/index.ts
@@ -1,1 +1,4 @@
 export * from './ApplicationLauncher';
+export * from './ApplicationLauncherItem';
+export * from './ApplicationLauncherGroup';
+export * from './ApplicationLauncherSeparator';