+++ packages/patternfly-4/react-core/src/components/Dropdown/Item.d.ts
@@ -1,9 +1,0 @@
-import { FunctionComponent } from 'react';
-import { DropdownItemProps } from './DropdownItem';
-
-export interface ItemProps extends DropdownItemProps {
-}
-
-declare const Item: FunctionComponent<ItemProps>;
-
-export default Item;