+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.23...@patternfly/react-docs@4.8.24) (2019-06-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.22...@patternfly/react-docs@4.8.23) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.21...@patternfly/react-docs@4.8.22) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.20...@patternfly/react-docs@4.8.21) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.19...@patternfly/react-docs@4.8.20) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.8.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.18...@patternfly/react-docs@4.8.19) (2019-06-19)
 
 **Note:** Version bump only for package @patternfly/react-docs