+++ packages/patternfly-4/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.8.19...@patternfly/react-inline-edit-extension@2.8.20) (2019-06-21)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.8.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.8.18...@patternfly/react-inline-edit-extension@2.8.19) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.8.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.8.17...@patternfly/react-inline-edit-extension@2.8.18) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.8.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.8.16...@patternfly/react-inline-edit-extension@2.8.17) (2019-06-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [2.8.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.8.15...@patternfly/react-inline-edit-extension@2.8.16) (2019-06-19)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension